Implemented a builder for the meal plan output data. Also fixed checkstyle issues in the User entity.
