+++ src/main/java/entities/User.java
@@ -63,12 +63,12 @@ public class User {
         return mealPlans;
     }
 
+    /**
+     * Saves the specified meal plan to the user's list of meal plans.
+     * @param mealPlan the meal plan to be saved
+     */
     public void saveMealPlan(MealPlan mealPlan) {
         mealPlans.add(mealPlan);
-    }
-
-    public void setGroceryList(GroceryList groceryList) {
-        this.groceryList = groceryList;
     }
 
     public void setGroceryList(GroceryList groceryList) {
