+++ src/main/java/use_case/load_meal_plan/LoadMealPlanInteractor.java
@@ -1,10 +1,7 @@
 package use_case.load_meal_plan;
 
-import java.util.ArrayList;
 import java.util.List;
-import java.util.Map;
 
-import entities.Ingredient;
 import entities.MealPlan;
 import entities.Recipe;
 import use_case.meal_plan.MealPlanOutputData;
@@ -13,7 +10,6 @@ import use_case.meal_plan.MealPlanOutputData;
  * Interactor for the Load Meal Plan Use Case.
  */
 public class LoadMealPlanInteractor implements LoadMealPlanInputBoundary {
-    private static final int MEAL_PLAN_SIZE = 3;
     private final LoadMealPlanDataAccessInterface dataAccessObject;
     private final LoadMealPlanOutputBoundary loadMealPlanPresenter;
 
@@ -30,40 +26,18 @@ public class LoadMealPlanInteractor implements LoadMealPlanInputBoundary {
     public void execute(LoadMealPlanInputData loadMealPlanInputData) {
         final int index = loadMealPlanInputData.getIndex();
         final List<MealPlan> mealPlans = dataAccessObject.getMealPlans();
-        final String[] recipeNames = new String[MEAL_PLAN_SIZE];
-        final String[] recipeImages = new String[MEAL_PLAN_SIZE];
-        final List<List<String[]>> recipeIngredients = new ArrayList<>();
-        final List<Map<String, Double>> recipeNutritionalValues = new ArrayList<>();
 
+        // Gathers the meal plan's recipes
         final MealPlan mealPlan = mealPlans.get(index);
-        int i = 0;
-        for (Recipe recipe : mealPlan.getRecipes()) {
-            recipeNames[i] = recipe.getRecipeName();
-            recipeImages[i] = recipe.getRecipeImage();
-            recipeIngredients.add(toOrderedString(recipe.getIngredients()));
-            recipeNutritionalValues.add(recipe.getNutritionalValues());
-            i++;
+        final List<Recipe> mealPlanRecipes = mealPlan.getRecipes();
 
-        }
-        final MealPlanOutputData mealPlanOutputData = new MealPlanOutputData(recipeNames, recipeImages,
-                recipeIngredients, recipeNutritionalValues);
+        // Builds the output data and sends it to the presenter
+        final MealPlanOutputData mealPlanOutputData = new MealPlanOutputData.Builder()
+                .buildRecipeNames(mealPlanRecipes)
+                .buildRecipeImages(mealPlanRecipes)
+                .buildRecipeIngredients(mealPlanRecipes)
+                .buildRecipeNutritionalValues(mealPlanRecipes)
+                .build();
         loadMealPlanPresenter.prepareSuccessView(mealPlanOutputData);
-    }
-
-    /**
-     * Converts a list of ingredients into a list of string arrays.
-     * @param ingredients the list of ingredients to be converted
-     * @return the list string array representation of the ingredients
-     */
-    private static List<String[]> toOrderedString(List<Ingredient> ingredients) {
-        final List<String[]> result = new ArrayList<>();
-        for (Ingredient ingredient : ingredients) {
-            final String[] ingredientEntry = new String[MEAL_PLAN_SIZE];
-            ingredientEntry[0] = ingredient.getName();
-            ingredientEntry[1] = ingredient.getQuantity() + "";
-            ingredientEntry[2] = ingredient.getUnit();
-            result.add(ingredientEntry);
-        }
-        return result;
     }
 }
