+++ src/test/java/use_case/meal_plan/MealPlanInteractorTests.java
@@ -10,7 +10,6 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 
-import static org.bson.assertions.Assertions.fail;
 import static org.junit.jupiter.api.Assertions.*;
 
 public class MealPlanInteractorTests {
